Capture all selected errors
This allows us to track more errors than before. Now we can also track
`SyntaxError`s when a file is required which has a syntax error during a
transaction. When a file is being required that doesn't exist we capture
a `LoadError` and we can also track the `NotImplementedError` for
methods and/or features that haven't been implemented yet.